refactor(BaseballGame): BaseballGame의 유틸 메소드 분리
- utils.js: generateRandomNumbers, compareNumbers, generateResultString
